Fix tower camera invincibility inconsistencies
If you have invincibility enabled, the tower camera behavior is
inconsistent.

In ascending towers, you can "push" the camera upwards; however you
cannot push it downwards; at least it stays still when it comes up to
you if you stay still. In descending towers, the camera moves quicker
when you're at the bottom of the screen, but it's slower than your
falling speed and quickly loses sight of you; the camera can be pushed
upwards; unfortunately it also does a "bumping" motion if you're
standing still when the camera reaches you, which gets real annoying and
isn't particularly pleasant to look at.

There are two problems, so this does two fixes:

1. Pushing the camera now applies the appropriate counter-offset
   depending on the direction of the tower. You can now push the camera
   downwards in ascending towers.

2. To fix the "bumping" when the camera reaches you if you stand still,
   there are now a 8-pixel-high "gray areas" at the top and bottom of
   the screen where the camera simply won't move if you're in them.

Doing these camera offsets instead of simply canceling the movement if
the player is offscreen is a bit ugly... but it works for now.